Move _override_es_config to BaseService
